lib/tempest: add wait for Glance image import
Glance image import is asynchronous and may be configured to do image
conversion.  If image import is being used, it's possible that the
tempest configuration code is executed before the import has
completed and there may be no active images yet.  Introduce a
tempest configuration option, TEMPEST_GLANCE_IMPORT_WAIT_TIME that
will specify the number of seconds we'll wait before trying again
to see if there are any active images.  (Default value is 12.)

You can see an example of this happening in real life:
https://review.opendev.org/c/openstack/glance/+/841278/1#message-456096e48b28e5b866deb8bf53e9258ee08219a0

Change-Id: Ie99f12691d9062611a8930accfa14d9540970cc5