Update patch set 3
Patch Set 3:

Dr. Jens Harbott wrote:

> The answer to the last question should be the same as the answer proposed here.

With my zuul-jobs reviewer hat on, I'm not sure having a generic role hard-code an assumption (or even default to it) that systems should try resolving opendev.org is a good idea. At a minimum, it would need some escape hatch for jobs being run in environments where nodes may not have access to a fully open recursive resolver, either to skip the check or substitute an alternative record which is known to be resolvable by those nodes... but beyond that, should the fips-setup role necessarily even depend on having working DNS resolution?

Patch-set: 3