Convert dbcounter to legacy setup.py
For some reason on Jammy with fully-installed constraints, the pip
"new way" of installing dbcounter fails with errors in its own
library for unknown reasons. This is probably the only package we
are installing that uses the "new way" so it breaks. There's no
reason for this tiny plugin to use the "new way" especially since it
requires more files and work, so just convert this back to setup.py
to sidestep the problem.

Change-Id: I1254a76854f43f74c62fad0dddb6c11d51367eae