Pin Tempest to 26.1.0 tag for stable/victoria testing
Stable victoria is in Extended maintenance state[1] and
we need to pin tempest in stable/victoria testing.

We use Tempest 26.1.0 because of oslo.utils version bump in Tempest
27.0.0 and 28.0.0 which is not compatible with stable/victoria
upper constraints so these versions do not work in stable/victoria.

Tempest 26.1.0 version is compatible for victoria testing from tempest
side as well as from upper constraints side.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/843293
Change-Id: I3f6c3914d0a35155dc27bbaf6e8db92b856a4f1f