DNM Using this change to test the depends on before we merge it
Devstack is a major user of the stage-output role. We've discovered the
way the stage output role loops over file processing is unfortunately
quite slow and has led to devstack based jobs timing out. To address
this I've rewritten the slow bits of that role into an ansible module
which should be significantly quicker. To sanity check it works for
devstack we have this change.

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/855402
Change-Id: I26e057106f2081e03d66886e969b037f1f203171