Add IMG_IMAGES_MIN_SIZE control
This adds a "minimum size" control for the image that we fetch and
upload to glance. This can be used to inflate the size of smaller
images (like cirros) to something more typical for the purposes of
increasing the realism of image handling. We use truncate to do this,
so the image on disk (before upload) is sparse and takes up no more
space than before.

Change-Id: Ia48e80b326f960fde3709a3813046a1ef2fe5cda