Rename lib/neutron-legacy to lib/neutron
Previously lib/neutron-legacy module was old module and the plan was to
remove it and use new lib/neutron instead. But as new lib/neutron module
was never really finished and used, it was decided to remove it from
Devstack.
This patch renames lib/neutron-legacy to be lib/neutron as it is defacto
neutron module in Devstack.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/867065

Closes-Bug: #1996748

Change-Id: I3c42bc521bf8debf40700b559505cc30e2a38e1b