Glance: Properly copy rootwrap.d/* when USE_VENV is True
The glance_store rootwrap base directory is currently set to a location
that only exists in "regular" installations. When installing Glance in a
virtual environment (by setting USE_VENV to True), copy_rootwrap tries
to copy files from that location, which does not exist, and fails.

This patch properly sets the rootwrap base directory, allowing files to
be copied.

Closes-Bug: #1967268
Change-Id: Idd2b03cdfeea758b1d4c77ae4ea7ab8ebd349257