[Train Only] Run ensure-rust role
In train tempest is pinned to 26.1.0(requires paramkio>=2.7.0)
and since paramiko-3.0.0(requires cryptography>=3.3) release
jobs are failing as they are trying to compile cryptography
which needs rust to be installed or upgrade pip so
cryptography wheel binaries can be installed.

To setup rust ensure-rust role is executed in
pre run.

Partial-Bug: #2004490
Change-Id: I2ec1c34ec11f50fef06f71b4a01976aa1d92a6bb