Set dhcp_client based on cirros version
This change allows us to bump the default cirros version
in devstack. Since cirros version 0.6.1 dhcpcd is the default
dhcp client. The older cirros images used udhcpc client by default
which is also the default client in Tempest.
This patch makes devstack configure dhcpcd client in tempest.conf
if cirros >= 0.6.1 is going to be used in scenario tests.

Closes-Bug: #2007973
Change-Id: I2738c3b1d302c6656ce2c209671ea954fbc1b05b