Add SERVICE_REPORT_INTERVAL knob
Heavily-loaded workers in CI consistently fail to complete the
service checkin task, which is configured for every ten seconds in
nova and cinder. This generates additional load on the database server
as well as consumes a threadpool worker. If we're not making the
deadline, there's really no point in having it be so high. Further,
since the workers must remain up for all the tempest tests we're
running against them, there's really no benefit to a fast-fail
detection.

This adds a knob that, if set, controls the report_interval for
nova and cinder, and sets service_down_time to 6x that value, which is
consistent with the default scale.

Change-Id: Idd7aa1daf354256b143a3778f161cfc72b318ea5