add support for zswap and ksmtuned
This change add a new lib/host-mem file and moves the existing
ksm support to a new confiugre_ksm funtion.
Addtional supprot for ksmtuned is added with a new flag
"ENABLE_KSMTUNED" which default to true.

This change also adds support for zswap. zswap is disabled
by default. when enabled on ubuntu lz4 will
be used as the default compressor and z3fold as the zpool.
on non debian distos the compressor and zpool are not set.
the default values should result in very low overhead although
the zstd compressor may provide better overall perfroamce in ci
or with slow io due to the higher compression ratio.

addtionally memory and network sycctl tunnign are optoinally applied
to defer writes, prefer swapping and optimise tcp connection
startup and keepalive. The sysctl tunings are disabled by default
The base devstack job has been modifed to enable zram and sysctl
tuning.

Both ksm and zswap are wrapped by a tune_host function
which is now called very early in devstack to ensure
they are configured before any memory/network intensive
operations are executed.

Change-Id: Ia5202d5a9903492a4c18b50ea8d12bd91cc9f135