[DNM] testing centos9/rocky9 jobs with global venv enabled
It appears this is caused by a change in pip after 21.1.3
i have not found the specific version but it is broken in
the latest 23.x.y

the delta in behavior apears to be in newer pip wheel is
venderod or build into the pip tool
that cause an editiabel wheel to be built instead of useing
egg format. i belive that cause pbr to not generate
the wsgi files.

pinning the pip version locally appears to resolve the
issues temporaly so lets see if the ci agrees

Change-Id: Ife6c97869d50cf59b595a5c7477d2d985b9cca38