Handle empty clouds.yaml files in update_clouds_yaml
If the clouds.yaml file ends up empty for some reason (accidental
overwrites by users for example) then we load the file as None.
Unfortunately None doesn't act like a default and we will throw
exception latest in this case.

Handle this situation by checking for a None value and reseting to an
empty dict in this situation.

Change-Id: I7d7fe8dbf38f4cc4463e0620a7c316cac7e52053