Enable GLOBAL_VENV=True for all distros
This change removes the workaround for centos 9 stream
and related distos which disabled the use of the GLOBAL_VENV
due to a lack fo supprot for PEP-518 in openstack projects.

PEP-518: https://peps.python.org/pep-0518/
"Specifying Minimum Build System Requirements for Python Projects"
defines the pyproject.toml format and file.

pip 20.2 deprecated support for calling "setup.py install" when a
python project did not support PEP-518 as part of adding supprot for
PEP-517

PEP-517: https://peps.python.org/pep-0517/
"A build-system independent format for source trees"

In pip 23.1 the fallback to calling "setup.py install" was removed
and pip now vendors a default pyproject.toml to use when one is not
found.

This breaks our use of pbr to generate wsgi scripts and was the reason
GLOBAL_VENV=True failed on centos9 and related distros

This change depend on the changes to the openstack projects to enable
support for PEP: 518, 517 and 660

PEP-660: https://peps.python.org/pep-0660/
"Editable installs for pyproject.toml based builds (wheel based)"

PEP-660: will be supported in a new release of pbr however a fallback
exists in setutool >= 64.0.0 that means we do not strictly need to wait
for the PBR release provided the system has a new enough setuptools.

Closes-Bug: #2031639
Depends-On: https://review.opendev.org/c/openstack/keystone/+/899519
Depends-On: https://review.opendev.org/c/openstack/nova/+/899753
Depends-On: https://review.opendev.org/c/openstack/placement/+/899947
Depends-On: https://review.opendev.org/c/openstack/cinder/+/899952
Depends-On: https://review.opendev.org/c/openstack/glance/+/899951
Depends-On: https://review.opendev.org/c/openstack/neutron/+/899956
Depends-On: https://review.opendev.org/c/openstack/horizon/+/899953
Depends-On: https://review.opendev.org/c/openstack/ironic/+/927544
Depends-On: https://review.opendev.org/c/openstack/swift/+/899958
Change-Id: Ife6c97869d50cf59b595a5c7477d2d985b9cca38