Use only is_service_enabled n-novnc for console proxy config
Long ago, the configuration of VNC was not separated between the
controller node vs a compute node, so config for both was included on
every node of a multinode configured job despite the fact that the
nodes could be running (or not running) different services. The
NOVA_VNC_ENABLED environment variable is available as a way to indicate
that VNC should be configured on a compute node that is not running a
n-novnc service.

Later on, we separated the configs between controller vs compute node
to enhance clarity and understanding. Since then, we know that the
configure_console_proxies function configures only the proxy, not the
compute node -- so we don't actually need to consider NOVA_VNC_ENABLED
because the proxy will always have the n-novnc service enabled.

Currently, because we do consider NOVA_VNC_ENABLED when configuring a
console proxy, the nova.conf or nova_cellN.conf  (not the
nova-cpu.conf) on the subnode of a multinode job that is not running
n-novnc will end up containing console proxy settings. This is not
harmful but it could cause confusion for those debugging and looking
through logs and configs.

This removes consideration of NOVA_VNC_ENABLED when configuring a
console proxy to tidy things up a bit and reduce potential confusion.

Change-Id: I89d935f56c264e693b6422dbefc084f03d56449f