[DNM] Debug osc keyring issue #2042943
Jobs fails randomly since the keyring is enabled
for osc commands.
Due to parallel executions and due to some race
something is adding duplicate cache_ids in the keystore
and that leads to issues, and until the keystore file
is removed all further execution of osc commands fails.
Gather the keystore data when the command fails.

Related-Bug: #2042943
Change-Id: Idfdc7cb119ce81cc4a2d997cd60f3fabef0d05df