Stable-only: Force v1beta etcd api
Currently skip level upgrade job is broken because cinder receives
Not Found error when connecting to etcd3. This is because the recent
change in tooz[1] switched the default api version to v3, but during
the upgrade process etcd is not upgraded timely and still uses v3.3
which does not yet support v3 API.

This forces usage of v3beta, which is available in v3.3 as well as
v3.5 to avoid breakage after only cinder and tooz are updated. v1beta
is supported both by 3.3(in stable/2023.1) and 3.4(stable/2023.2 and
master).

[1] 738931d026aedb846ed3970cf642a6c570a2bceb

Change-Id: I07b391ce7641c3bc0563bdbad74e0378bc0856bb