Drop devstack-single-node-opensuse-15 nodeset
This nodeset was used by tempest-full-py3-opensuse15 which is being
removed in the depends on change.

This removal is happening because the OpenDev team is planning to remove
OpenSUSE LEAP 15 images as our node builds and mirrors are for 15.2
which is ancient and no one is currently working to modernize these test
environments. On top of that LEAP is apparently going away in the future
and will be replaced with another distro.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/909767
Change-Id: I361f2b3df8e3427542e7bc72168531edd1c33227