Add eBPF bcc tool profiling
Newer Linux kernels enable the use of eBPF to profile actions taken by
the system. There is a suite of eBPF profiling tools called bcc [0] that
appear particularly useful to profiling devstack and tempest runs.
Install this toolchain and add some swap and oomkiller profiling. This
should serve as a framework to add many more of these bcc tools to
devstack (there are many) as many appear potentially useful.

[0] https://github.com/iovisor/bcc

Change-Id: I65f234b18047e23c1d19b251816f1f581f8a866b