Drop devstack-gate roles from devstack zuul jobs
Devstack-gate has been retired and relying on a retired repo for roles
in zuul produces errors (because the roles are no longer in the HEAD of
the repo). It doesn't look like devstack actually relies on these roles
in d-g (if it does they can be ported into devstack instead) so we just
drop the roles specifier for d-g in the devstack job.

NOTE(elod.illes): nova-ceph-multistore job is broken in this branch, so
it is set as non-voting in check queue and removed from gate queue
until it gets fixed.

NOTE(elod.illes): grenade is set as non-voting as it won't pass until
grenade fix [1] is not landed, which is a 'catch 22' as it won't pass
until this patch is not landed.

[1] https://review.opendev.org/c/openstack/grenade/+/919818

Change-Id: I28a39b31f71153a602d41cefe621a216d09a290f
(cherry picked from commit d487302e60bf295122ab04ba5d3b799132ed0b17)
(cherry picked from commit 503116557a0b5ed91aa7bea2ef6eaa967ce15712)