Ensure $DATA_DIR permissions after etc3 install
ETCD_DATA_DIR uses the DATA_DIR value as base location. There are
some scenarios where the creation of this folder modifies the
original permissions of the DATA_DIR folder and after the subsequent
configuration interruption the provisioning process. This change ensures
the original permissions are kept after the ETCD_DATA_DIR creation

Change-Id: I8fd1de0b3acc253a72ba1b2bc5f63ea389ecaa42