Do not use 'admin' role in service users
Service users are meant to be used in service-to-service
interaction, which should use 'service' ([keystone_authtoken].
service_token_roles) role only. Currently, nova and placement
service users are assigned 'admin' role also. There might
be changes required on services side to allow service interaction
with 'service' role only. To proceed/test the work/fixes, we
need to prepare a job which does not set admin role on service
user. For that job configuration, adding a variable here which
can be used to assign the admin role or not.

Change-Id: If4a2a6688b8fed09144cdc4bdf1329361ca2b524