Fixes the repo configs for Rocky9 and stream9
This PR fixes the repository configuration for OpenStack packages on Rocky
Linux 9 and CentOS Stream9. DevStack currently installs the repository
configuration from the `rdoproject.org/repos/openstack-xxx.rpm` package, but
this package is returning a 404 error and cannot be installed. Instead,
we will use the `centos-release-openstack-xxx package` from the CRB repository.

Change-Id: I1eb7468395c8d450ddc16151237b1cbc0c3636ea
Closes-Bug: #2075203