Fix error when wget is actually wget2
Some distributions provide "wget2" as "wget". Right now this seems to
only be Fedora 40, which is not supposed to work with Devstack, but this
very well might end up in CentOS at some point.

The behaviour of wget2 is a bit different from that of wget: it writes
more info on stdout and does not support all progress indicators. This
causes the get_extra_file function to print out wget2 messages as well
as the filename. This causes the installation of etcd3 to fail, since
the content of the etcd_file variable becomes:

    Progress type '\''dot:giga'\'' ignored. It is not implemented yet
    [0] Downloading ...
    <redacted>
    /home/vagrant/devstack/files/etcd-v3.4.27-linux-amd64.tar.gz'

We work around the issue by running wget in quiet mode when it is
actually wget2 in disguise.

Change-Id: Iac56c5fdc44caafbe9e7221151874f37f1c6f101