lib/nova: Allow specifying NOVA_SERVICE_DOWN_TIME
Currently we don't have a dedicated variable for the
[DEFAULT]service_down_time config option in Nova. Instead, it is
calculated by multiplying NOVA_SERVICE_REPORT_INTERVAL * 6.

This can be quite limiting depending on the CI environment, especially
third-party CI. While it is possible to set the necessary Nova config
options in [post-config], it would be a lot easier to be able to use
NOVA_SERVICE_DOWN_TIME.

This adds a new variable NOVA_SERVICE_DOWN_TIME which defaults to
NOVA_SERVICE_REPORT_INTERVAL * 6 if not specified, to retain existing
behavior as a default.

Change-Id: I514493139ee4e2d167cec9a2258a3b472b61a701