ovsdbserver loosing OVN_DBS_LOG_LEVEL custom config
During the startup process, OVN services are restarted multiple
times. Initially, the ovsdbserver log level is correctly configured
when a custom configuration value, OVN_DBS_LOG_LEVEL, is defined.
However, after a restart, it loses this value and reverts to the
default INFO level.

This patch ensures that the custom configuration is reapplied to
OVN services each time they are restarted.

Closes-bug: #2091019
Change-Id: Icba271292830204da94aa3353e93d52088d82eec