ovn_agent: start OVS processes with retries
Sometimes `ovsdb-server` takes a couple of restarts to be
really functional, and in some environments we have observed this kind
of error:

```
Job for ovsdb-server.service failed because a timeout was exceeded.
```

Systemd would automatically restart the service and eventually it'll be
functional.

Adding this retry will help to not have devstack failures in these
environments. We'll basically try a restart during $SERVICE_TIMEOUT with
15 seconds sleeps between attempts. This will relax how we check whether
OVS is started and whether it's functional.

Change-Id: I4133809466f1dd626b35fb82c88e46bd1fff06cb