Do not install suggested/recommended packages
apt by default attempts to install packages which are not actually
requires. This results in pulling unnecessary packages which may
conflict with the ones we aim to install virtualenv.

One example is some oslo packages installed when novnc is installed,
which are pulled via python3-nova listed as a suggested dependency.

Related-Bug: #2109592
Change-Id: Ib96f1762632eeed8ab751793d3eca7926c3c3873