[Neutron] Add TARGET_ENABLE_OVN_AGENT variable to enable OVN agent
The new flag ``TARGET_ENABLE_OVN_AGENT`` will be disabled by default.
If enabled:
* The OVN agent will be enabled, regardless of not being configured.
* The OVN Metadata agent will be disabled, regardless of being
  configured.

This variable will be used, initially, in the grenade jobs. It will
be used to test the migration from the OVN Metadata agent to the
OVN agent.

This variable will be removed in 2026.2, along with the OVN Metadata
agent, that is set as deprecated and marked for removal.

Related-Bug: #2112313
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I8f91e1cb8543da489f495b8cf5196e606a0f5eea