Add Swift storage policy to extend testing
Swift supports storage policies for a long time, making it possible to
store specific objects on different devices[1].

This patch adds a second storage policy to support testing API features
that rely on storage policies, eg. per-policy quotas[2]. There are no
additional Swift devices created and the existing ones are re-used to
keep the overhead to the minimum.

[1] https://docs.openstack.org/swift/latest/overview_policies.html
[2] https://github.com/openstack/swift/commit/cbba65ac

Signed-off-by: Christian Schwede <cschwede@redhat.com>
Change-Id: I6b1a35dbf1490b25e103f008eb058f406f0a24ea