Set password regex for security compliance
This patch enables use of the security_compliance.password_regex option
in keystone.conf by setting the $KEYSTONE_PASSWORD_REGEX variable.

The default value is the sample regex from the Security Compliance
section of the admin config guide. [1]

The patch that depends on this change adds the tempest option for the
user password minimum length, which is set in tempest.conf here using
the $KEYSTONE_PASSWORD_REGEX_MIN_LENGTH variable.

[1] https://docs.openstack.org/keystone/latest/admin/configuration.html#configuring-password-strength-requirements

Needed-By: https://review.opendev.org/c/openstack/tempest/+/954029
Signed-off-by: Douglas Mendizábal <dmendiza@redhat.com>
Co-Authored-By: Milana Levy <millevy@redhat.com>
Change-Id: I8a74b871227da1b40f8952f67b3a8e92a424bc0c