Use novnc from source by default
This change restores the default devstack behavior in the zuul jobs
by removing the override of NOVNC_FROM_PACKAGE from devstack-base.

When installed locally, devstack defaults to installing novnc from
git. As reported in bug #2109592, Ubuntu and possibly other distros
have a packaging bug where the python3-novnc package 1) exists and
2) depends on `oslo.config` and, as a result, `oslo.utils`. The
reason python3-novnc existing is a bug is that novnc has not had any
Python deliverable since the 0.6.0 release around 2016. So this
package is no longer used and is effectively empty since novnc fully
moved to using JavaScript.

For unrelated reasons, devstack creates the global venv with
--site-packages to install `libvirt-python`, which also means that
any other Python dependencies installed at the system level also
infect the devstack venv.

In the past, this was not a problem, but as of epoxy, Nova requires
a newer version of oslo than Ubuntu provides in the distro package.
This is where the python3-novnc package and its incorrect dependency
on oslo breaks CI.

This is not seen locally, as devstack uses novnc from git. This
change makes CI do that also.

Closes-Bug: #2109592

Change-Id: I8f018e1e57e3f54997d2cf55b1b3aa728e82899b
Signed-off-by: Sean Mooney <work@seanmooney.info>